requirement.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
Django==3.1
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.6
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.18.4
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.22
wrapt==1.12.1